requirements.txt
robotframework>=6.0